requirements.txt
playwright==1.44.0
psycopg2-binary==2.9.9
requests==2.31.0
